requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Testing frameworks
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
mypy==1.7.1

# Debugging
ipython==8.18.1
pdb-plus==1.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Performance testing
locust==2.17.0
django-silk==5.0.4

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# All production dependencies
-r requirements.txt
